Move ui routes to pages.go
